Run clean and lint again
There’s a block at the bottom of the CSS that doesn’t apply directly to the unprocessed code, but is there to reset styles against the HTML that’s inserted during the Kobo/Kindle processing step. This shows up as unused in lint, which is true, but is necessary for the final product.